bugfix: tiles were slotted into transpose spots
should have been b->tiles[j][i] = tile_ctor(i, j)
because of the way that indices vs x and y
coords works. Since the i coordinate is the
row, and j is the column, we need j to correspond
to the x and the i to correspond to the y.
Unfortunately, in the board init, this was
done the wrong way, but was done correctly everywhere
else.